+++ examples/pairwise_rmsd.py
@@ -2,7 +2,7 @@ import numpy as np
 from pytraj.base import *
 from time import time
 
-TRAJ = TrajReadOnly()
+TRAJ = TrajectoryIterator()
 TRAJ.top = Topology("../tests/data/Tc5b.top")
 TRAJ.load("../tests/data/md1_prod.Tc5b.x")